some method on repo having their own class, update on the templates, and Controllers are seperated.
