+++ src/main/java/id/ac/ui/cs/advprog/eshop/controller/HomeController.java
@@ -1,0 +1,12 @@
+package id.ac.ui.cs.advprog.eshop.controller;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+
+@Controller
+@RequestMapping("/")
+class HomeController {
+    @GetMapping
+    public String homePage() {
+        return "Homepage";
+    }
+}
