+++ src/main/resources/templates/CarList.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8">
     <title>Car List</title>
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-</head>
+    <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro&display=swap" rel="stylesheet">
+  </head>
 <body>
 
     <nav class="navbar navbar-expand-lg navbar-dark" style="background-color: #0b806c; font-family: 'Source Sans Pro', sans-serif;">
