+++ src/main/resources/templates/CreateProduct.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8">
     <title>Create New Product</title>
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"rel="stylesheet" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERd`knLPMO" crossorigin="anonymous">
-</head>
+    <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro&display=swap" rel="stylesheet">
+  </head>
 <body>
 
     <nav class="navbar navbar-expand-lg navbar-dark" style="background-color: #0b806c; font-family: 'Source Sans Pro', sans-serif;">
