+++ src/test/java/id/ac/ui/cs/advprog/eshop/repository/ProductRepositoryTest.java
@@ -53,18 +53,4 @@ class ProductRepositoryTest {
         assertFalse(productIterator.hasNext());
         productRepository.clear();
     }
-
-    @Test
-    void testDelete() {
-        Product product = new Product();
-        product.setProductId("eb558e9f-1c39-460e-8860-71af6af63bd6");
-        product.setProductName("Sampo Cap Bambang");
-        product.setProductQuantity(100);
-        productRepository.create(product);
-
-        productRepository.delete("Sampo Cap Bambang");
-        Iterator<Product> productIterator = productRepository.findAll();
-        assertFalse(productIterator.hasNext());
-        productRepository.clear();
-    }
 }
